delete playbook_service_provider.dart; conditions all file that depend on it
