+++ lib/view/components.dart
@@ -1,8 +1,8 @@
+import 'package:BookVerse/helper/book_authors.dart';
 import 'package:BookVerse/helper/push_navigation.dart';
 import 'package:BookVerse/view/pages/detail_page.dart';
 import 'package:flutter/material.dart';
 import 'package:BookVerse/model/book_model.dart';
-import 'package:BookVerse/provider/playbook_services_provider.dart';
 
 Widget bookGridTile({
   required Book book,
