+++ lib/view/pages/detail_page.dart
@@ -1,9 +1,9 @@
 import 'dart:developer';
 
+import 'package:BookVerse/helper/book_authors.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:BookVerse/model/book_model.dart';
-import 'package:BookVerse/provider/playbook_services_provider.dart';
 import 'package:BookVerse/provider/search_provider.dart';
 import 'package:BookVerse/provider/bookmark_provider.dart';
